pyproject.toml
[project]
name = "glyph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firecrawl>=0.1.0",
    "langchain-openai>=0.3.16",
    "langgraph>=0.0.15",
    "markdown>=3.5.2",
    "python-dotenv>=1.0.0",
    "ruff>=0.11.8",
    "streamlit>=1.32.0",
    "tavily-python>=0.3.1",
]
